+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.4.0
+
++ Added `StoreEvent` -> `PutValue`,`ValueLoaded`,`ValueSaved`,`ValueLoadError`,`ValueSaveError`
++ Added `SpecialMethodsLogic` -> `putAndWriteAll`,`putAndWriteAllSync`
++ Changed  `void put(String key, dynamic value);`to `ICFBStore put(String key, dynamic value);`
+
+
 # 0.3.0
 - Added `getMap`